+++ src/modules/auth/ButtonWithAuthz.js
@@ -12,6 +12,9 @@ import { useTranslation } from 'react-i18next';
 	* resources
 		* the list of resources from the userinfo
 
+	* hideOnUnauthorizedAccess
+		* the option to hide the button completely instead of disabling it, default is false
+
 	* Language localisations for generic `unauthorized message` can be added to the translation.json files of
 	  public/locales/en & public/locales/cs of the product where component ButtonWithAuthz is used. The default
 	  message is `You do not have rights` and it can be re-set in locales as e.g.
@@ -41,6 +44,7 @@ import { useTranslation } from 'react-i18next';
 				onClick={() => { deleteApp() }}
 				resource="myapp:mylist:write"
 				resources={resources}
+				hideOnUnauthorizedAccess={true}
 				>
 				<span className="cil-trash pr-2" />
 				{t('MyApp|Delete')}
@@ -52,24 +56,32 @@ import { useTranslation } from 'react-i18next';
 
 export function ButtonWithAuthz(props) {
 	const { t, i18n } = useTranslation();
-	let disabled = props.resources ? props.resources.indexOf(props.resource) == -1 : true;
-	let title = props.title;
+	let childProps = {...props}; // Create a new child component to have option to remove props
+	let disabled = childProps.resources ? childProps.resources.indexOf(childProps.resource) == -1 : true;
+	// If defined, hide the disabled button
+	let hide = childProps.hideOnUnauthorizedAccess ? true : false;
+	// Remove hideOnUnauthorized element from props to avoid react warnings
+	if (childProps.hideOnUnauthorizedAccess) {
+		delete childProps["hideOnUnauthorizedAccess"];
+	}
+	let title = childProps.title;
 	// Check on title eventually passed in the props
 	if (disabled) {
 		 title = t("You do not have access rights to perform this action");
 	}
 	// Check on disabled eventually passed in the props
-	if (props.disabled && disabled == false) {
-		disabled = props.disabled;
+	if (childProps.disabled && disabled == false) {
+		disabled = childProps.disabled;
 	}
 
 	return (
+		hide && disabled ? null :
 		<Button
-			{...props}
+			{...childProps}
 			title={title}
 			disabled={disabled}
 			>
-			{props.children}
+			{childProps.children}
 		</Button>
 	)
 }